Updating imports to reflect connection->_http module rename.
